opsagent/ui/pyproject.toml
[project]
name = "chatui-sso"
version = "0.1.0"
description = "DAPE OpsAgent Manager - Flask Chat UI with SSO"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "gunicorn>=21.2.0",
    "psycopg2-binary==2.9.9",
    "python-dotenv==1.0.0",
    "redis>=5.0.0",
    "azure-storage-blob>=12.19.0",
    "azure-identity>=1.15.0",
]
